Convert teamStatsSlice to typescript
