+++ store/teamStatsSlice.ts
@@ -1,0 +1,31 @@
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { ConferenceMap } from "@/schemas";
+
+interface TeamStatsState {
+  teamStats?: ConferenceMap;
+  conferenceArrays: {
+    [conferenceName: string]: string[];
+  };
+}
+
+const initialState: TeamStatsState = {
+  conferenceArrays: {},
+};
+
+const teamStatsSlice = createSlice({
+  name: "teamStats",
+  initialState,
+  reducers: {
+    addToStateFromDB(state, action: PayloadAction<ConferenceMap>) {
+      state.teamStats = { ...action.payload };
+    },
+    addConferenceArrays(state, action: PayloadAction<ConferenceMap>) {
+      Object.keys(action.payload).forEach((conf) => {
+        state.conferenceArrays[conf] = Object.keys(action.payload[conf]);
+      });
+    },
+  },
+});
+
+export const teamStatsActions = teamStatsSlice.actions;
+export default teamStatsSlice;
